Add load and 'save' functionality.
Load gets a prepared text file with a sample scribble and displays it on
top of the SVG file.

Save occurs whenever a path is created. The canvas is converted into
JSON and the output is placed in the textarea. On Rails, this will be a
hidden field which can then be POSTed to the server for saving to a
database.